# backend/internal/launchcode/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Facade Launch API
  version: 1.0.0
servers:
  - url: /
    description: Current LaunchServer origin
...
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Ant-Api-Key
      description: 应用设置中启用 Launch API Key 后必填
  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: string
        authHeader:
          type: string
          description: 鉴权失败时返回期望的 header 名
    Profile:
      type: object
      description: 实例完整快照
      properties:
        profileId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
        profileName:
          type: string
          example: buyer-001
        userDataDir:
          type: string
        coreId:
          type: string
        fingerprintArgs:
          type: array
          items:
            type: string
        proxyId:
          type: string
          example: proxy-us
        proxyConfig:
          type: string
          description: 原生代理链接（direct:// / http:// / https:// / socks5://）
        proxyBindName:
          type: string
        proxyBindUpdatedAt:
          type: string
        launchArgs:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
          example: [电商]
        keywords:
          type: array
          items:
            type: string
          example: [buyer-001]
        groupId:
          type: string
        launchCode:
          type: string
          example: BUYER_001
        running:
          type: boolean
        debugPort:
          type: integer
        debugReady:
          type: boolean
        pid:
          type: integer
        runtimeWarning:
          type: string
        lastError:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        deletedAt:
          type: string
        lastStartAt:
          type: string
        lastStopAt:
          type: string
...
    LaunchRequestParams:
      type: object
      description: 仅影响本次启动的临时参数（不持久化覆盖实例配置）
      properties:
        launchArgs:
          type: array
          items:
            type: string
          description: 本次启动临时附加参数
        startUrls:
          type: array
          items:
            type: string
          description: 本次启动额外打开的网址
        skipDefaultStartUrls:
          type: boolean
          description: 是否跳过实例默认启动 URL
        proxyId:
          type: string
          description: 仅本次启动使用的代理池节点
        proxyConfig:
          type: string
          description: 仅本次启动使用的自定义代理链接
    LaunchSelector:
      type: object
      description: |
        目标实例选择条件。新接入推荐使用本对象；
        也可把同名字段放在请求体顶层（兼容写法）。
      properties:
        code:
          type: string
          description: launchCode
          example: BUYER_001
        key:
          type: string
          description: 兼容字段，语义同 code
        profileId:
          type: string
        profileName:
          type: string
        keyword:
          type: string
        keywords:
          type: array
          items:
            type: string
        tag:
          type: string
        tags:
          type: array
          items:
            type: string
        groupId:
          type: string
        matchMode:
          type: string
          enum: [unique, first, all]
          description: |
            unique=必须唯一命中；first=取第一个；all=批量（仅 POST /api/launch）
          example: unique
...
    LaunchRequest:
      allOf:
        - $ref: '#/components/schemas/LaunchRequestParams'
        - type: object
          properties:
            selector:
              $ref: '#/components/schemas/LaunchSelector'
            code:
              type: string
              description: 兼容写法，等价于 selector.code
            key:
              type: string
            profileId:
              type: string
            profileName:
              type: string
            keyword:
              type: string
            keywords:
              type: array
              items:
                type: string
            tag:
              type: string
            tags:
              type: array
              items:
                type: string
            groupId:
              type: string
            matchMode:
              type: string
              enum: [unique, first, all]
    RuntimeRequest:
      type: object
      description: runtime 系列接口请求体（不支持 matchMode=all）
      properties:
        selector:
          $ref: '#/components/schemas/LaunchSelector'
        code:
          type: string
        key:
          type: string
        profileId:
          type: string
        profileName:
          type: string
        keyword:
          type: string
        keywords:
          type: array
          items:
            type: string
        tag:
          type: string
        tags:
          type: array
          items:
            type: string
        groupId:
          type: string
        matchMode:
          type: string
          enum: [unique, first]
    RuntimeSessionRequest:
      allOf:
        - $ref: '#/components/schemas/RuntimeRequest'
        - $ref: '#/components/schemas/LaunchRequestParams'
        - type: object
          properties:
            timeoutMs:
              type: integer
              description: 等待 debugReady 的超时（毫秒），默认 45000，范围 1000–120000
              example: 45000
...
    LaunchSuccessResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        pid:
          type: integer
        debugPort:
          type: integer
        debugReady:
          type: boolean
        runtimeWarning:
          type: string
        cdpPort:
          type: integer
          example: 9333
        cdpUrl:
          type: string
          example: http://127.0.0.1:9333
    LaunchBatchSuccessResponse:
      type: object
      description: matchMode=all 时的批量启动响应
      properties:
        ok:
          type: boolean
          example: true
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/LaunchSuccessResponse'
    ProfileRuntimeResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
        profileId:
          type: string
        profileName:
          type: string
        launchCode:
          type: string
        running:
          type: boolean
        pid:
          type: integer
        debugPort:
          type: integer
        debugReady:
          type: boolean
        runtimeWarning:
          type: string
        lastError:
          type: string
        lastStartAt:
          type: string
        lastStopAt:
          type: string
        cdpPort:
          type: integer
        cdpUrl:
          type: string
        profile:
          $ref: '#/components/schemas/Profile'
        stopped:
          type: boolean
          description: stop 接口成功时为 true
    RuntimeSessionResponse:
      allOf:
        - $ref: '#/components/schemas/ProfileRuntimeResponse'
        - type: object
          properties:
            ready:
              type: boolean
              description: true 表示可立即 attach
            timeoutMs:
              type: integer
            waitTimedOut:
              type: boolean
            retryable:
              type: boolean
...
paths:
...
  /api/launch:
    post:
      tags: [Launch]
      operationId: launchBySelector
      summary: 按 selector 启动
      description: |
        更灵活的启动入口。支持 `selector` 或兼容顶层选择字段，以及 `launchArgs` / `startUrls` / `skipDefaultStartUrls`。

        - selector 为空且无任何兼容顶层选择字段时返回 400
        - `matchMode=all` 只在本接口可用，返回批量结果
        - `proxyId` / `proxyConfig` 只影响本次启动，不覆盖实例原代理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LaunchRequest'
            examples:
              bySelector:
                summary: 按 keyword + tags
                value:
                  selector:
                    keyword: checkout
                    tags: [电商, 北美]
                    matchMode: unique
                  skipDefaultStartUrls: true
              byCodeCompat:
                summary: 顶层 code 兼容写法
                value:
                  code: BUYER_001
                  skipDefaultStartUrls: true
      responses:
        '200':
          description: 启动成功（单实例或批量）
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/LaunchSuccessResponse'
                  - $ref: '#/components/schemas/LaunchBatchSuccessResponse'
        '400':
          description: selector 缺失或请求体非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 未命中实例
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: selector 命中多个实例（unique 模式）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/runtime/session:
    post:
      tags: [Runtime]
      operationId: prepareRuntimeSession
      summary: 准备可接管会话
      description: |
        按 selector 命中实例，必要时自动启动，并在超时内等待 `debugReady=true`。
        就绪后用返回的 `cdpUrl` 直连该实例调试端口。

        | HTTP | 含义 |
        |------|------|
        | 200 + ready=true | 可直接 attach |
        | 202 + ready=false | 已处理但未 ready，可稍后重试 |

        不支持 `matchMode=all`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuntimeSessionRequest'
            example:
              selector:
                code: BUYER_001
              timeoutMs: 45000
              skipDefaultStartUrls: true
      responses:
        '200':
          description: 实例已 ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSessionResponse'
              example:
                ok: true
                ready: true
                waitTimedOut: false
                retryable: false
                profileId: 550e8400-e29b-41d4-a716-446655440000
                launchCode: BUYER_001
                running: true
                debugReady: true
                cdpPort: 9333
                cdpUrl: http://127.0.0.1:9333
                timeoutMs: 45000
        '202':
          description: 已处理但暂未 ready
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSessionResponse'
        '400':
          description: selector 缺失或 matchMode 非法
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: API Key 无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 目标实例不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...